requirements.txt
discord
git+https://github.com/Rapptz/discord.py
python-dotenv
pymongo